pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
[project]
name = "multitool-raiz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "distro==1.9.0",
    "pillow>=12.0.0",
    "psutil==7.1.2",
    "pywin32==311",
    "screeninfo==0.8.1",
]
[tool.setuptools.packages.find]
where = ["multitool_v2"]
